requirements.txt
fastapi>=0.100.0
uvicorn[standard]==0.27.0
python-dotenv>=1.0.0
google-cloud-aiplatform>=1.35.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
langgraph>=0.0.10
chromadb>=0.4.15
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart==0.0.6
python-docx==0.8.11
PyPDF2==3.0.1
numpy==1.24.3
faiss-cpu==1.7.4
tavily-python